reorganized structure of components
