+++ src/app/components/StateRoute/PerformanceVisx/BarGraphComparison.tsx
@@ -13,8 +13,8 @@ import { makeStyles } from '@material-ui/core/styles';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import FormControl from '@material-ui/core/FormControl';
-import { onHover, onHoverExit, deleteSeries, setCurrentTabInApp } from '../actions/actions';
-import { useStoreContext } from '../store';
+import { onHover, onHoverExit, deleteSeries, setCurrentTabInApp } from '../../actions/actions';
+import { useStoreContext } from '../../store';
 
 /* TYPESCRIPT */
