+++ src/app/components/StateRoute/PerformanceVisx/BarGraphComparisonActions.tsx
