+++ src/app/components/StateRoute/PerformanceVisx/RenderingFrequency.tsx
@@ -2,8 +2,8 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable react/prop-types */
 import React, { useState, useEffect } from 'react';
-import { onHover, onHoverExit, setCurrentTabInApp } from '../actions/actions';
-import { useStoreContext } from '../store';
+import { onHover, onHoverExit, setCurrentTabInApp } from '../../../actions/actions';
+import { useStoreContext } from '../../../store';
 
 const RenderingFrequency = props => {
   const perfData = props.data;
