+++ src/app/containers/StateContainer.tsx
@@ -9,7 +9,7 @@ import {
   NavLink,
   Switch,
 } from 'react-router-dom';
-import StateRoute from '../components/StateRoute';
+import StateRoute from '../components/StateRoute/StateRoute';
 import DiffRoute from '../components/DiffRoute';
 
 interface StateContainerProps {
